+++ packages/atomic/src/components/atomic-breadcrumb-manager/atomic-breadcrumb-manager.tsx
@@ -82,8 +82,9 @@ export class AtomicBreadcrumbManager implements InitializableComponent {
     return (
       <button
         part="breadcrumb-button"
-        class="text-on-background-variant breadcrumb-button flex items-center hover:text-primary-variant"
+        class="text-on-background-variant breadcrumb-button flex items-center hover:text-primary-variant w-full"
         aria-label={this.strings.breadcrumb({value})}
+        title={value}
         onClick={() =>
           this.breadcrumbManager.deselectBreadcrumb(breadcrumbValue)
         }
@@ -109,13 +110,15 @@ export class AtomicBreadcrumbManager implements InitializableComponent {
     field: string,
     children: VNode[]
   ) {
+    const label = this.bindings.store.state.facets[facetId].label || field;
     return (
       <div class="flex text-sm" part="breadcrumb-wrapper">
         <span
-          class="text-on-background mr-2 flex-shrink-0"
+          title={label}
+          class="breadcrumb-label text-on-background mr-2 flex-shrink-0 overflow-ellipsis overflow-hidden"
           part="breadcrumb-label"
         >
-          {this.bindings.store.state.facets[facetId].label || field}:
+          {label}:
         </span>
         {children}
       </div>