+++ packages/atomic/src/components/facets/facet-value/facet-value.tsx
@@ -27,10 +27,13 @@ export const FacetValue = (props: FacetValueProps) => {
       />
       <label
         htmlFor={id}
-        class="ml-3 flex flex-row text-on-background flex-grow cursor-pointer"
+        class="ml-3 text-on-background cursor-pointer flex-shrink whitespace-nowrap overflow-ellipsis overflow-hidden flex flex-auto"
+        title={props.label}
       >
-        <span class="my-auto">{props.label}</span>
-        <span class="ml-auto my-auto self-end text-on-background-variant">
+        <span class="whitespace-nowrap overflow-ellipsis overflow-hidden flex-auto">
+          {props.label}
+        </span>
+        <span class="text-on-background-variant flex-shrink-0">
           {props.numberOfResults}
         </span>
       </label>