+++ packages/headless/src/api/search/plan/plan-request.ts
@@ -2,10 +2,8 @@ import {
   BaseParam,
   ContextParam,
   LocaleParam,
-  PipelineParam,
-  QueryParam,
-  SearchHubParam,
-} from '../search-api-params';
+} from '../../platform-service-params';
+import {PipelineParam, QueryParam, SearchHubParam} from '../search-api-params';
 
 export type PlanRequest = BaseParam &
   SearchHubParam &