+++ packages/headless/src/api/search/search-api-client.ts
@@ -18,7 +18,7 @@ import {PlanRequest} from './plan/plan-request';
 import {QuerySuggestRequest} from './query-suggest/query-suggest-request';
 import {FacetSearchRequest} from './facet-search/facet-search-request';
 import {SearchAppState} from '../../state/search-app-state';
-import {BaseParam, baseSearchRequest} from './search-api-params';
+import {baseSearchRequest} from './search-api-params';
 import {RecommendationRequest} from './recommendation/recommendation-request';
 import {ProductRecommendationsRequest} from './product-recommendations/product-recommendations-request';
 import {Logger} from 'pino';
@@ -30,6 +30,9 @@ import {
 } from './search-api-client-middleware';
 import {PreprocessRequest} from '../preprocess-request';
 import {HtmlRequest} from './html/html-request';
+import {findEncoding} from './encoding-finder';
+import {TextDecoder} from 'web-encoding';
+import {BaseParam} from '../platform-service-params';
 
 export type AllSearchAPIResponse = Plan | Search | QuerySuggest;
 
@@ -185,7 +188,10 @@ export class SearchAPIClient {
       ...this.options,
     });
 
-    const body = await response.text();
+    const encoding = findEncoding(response);
+    const buffer = await response.arrayBuffer();
+    const decoder = new TextDecoder(encoding);
+    const body = decoder.decode(buffer);
 
     if (isSuccessHtmlResponse(body)) {
       return {success: body};